Sort out #map_sidebar padding and .big-green-banner overlap
The big-green-banner can no longer extend outside the sidebar
without causing a horizontal scroll bar. So we just set it to
100% width (with negative margins compensating for padding on
#map_sidebar parent). This also means $col_click_map_dark is
no longer required, since that little triangle is no longer
added to the banner.

The resizing of the banner also required some tweaking to the
size of the chevron-white SVG images.