Replace overhanging report-a-problem-sidebar with wider map_sidebar
Since the map_sidebar now scrolls by itself, elements can no
longer overhang its right edge. Very few cobrands include an
overhanging element, but for the ones that do (eg: Zurich)
we now have a .with-notes variant of the map_sidebar that
adds a second column to the sidebar, perfect for containing
secondary notes about the form.

We add and remove the .with-notes class automatically
based on the presence of a #report-a-problem-sidebar.

Annoyingly the .with-notes class needs to be added to both
map_box and map_sidebar (since the size of both needs to be
adjusted to make room for the extra column).